Calculate RequestGroup resource provider mapping
If the port, provided in the server create request, has requested
resources then nova needs to communicate the resource allocation it
made during the scheduling to neutron.

As a single neutron port's resource request is included in the
allocation candidate query as a separate request group therefore the
alloction for a port will come from a single RP. Neutron expects that
this single RP uuid is communicated during the binding of the port in
the binding_profile.

Each Neutron port's resource request is represented as a RequestGroup in
the RequestSpec.

This patch introduces a code in RequestSpec to map RequestGroups to
resource providers in a generic way, base on the resource requests and
required traits in the groups, the overall allocation made for the
RequestSpec and the provider summary of the RPs in the allocation.

Subsequent patches will integrate this code into the server create code
path to create the mapping, then change the port binding code path to
use this mapping to communicate the resource provider of the port in the
binding profile.

blueprint bandwidth-resource-provider

Change-Id: Id7827fe8dc27112e342dc25c902c8dbc25f63b94