Send RP uuid in the port binding
The neturonv2/api already receives information about which resource
provider fullfills the resource needs of each neutron port. So in this
patch we include this infomation in the binding:profile of the neutron
port during the port binding.

Also make sure that this data is removed from the binding:profile when
the port is unbound.

blueprint bandwidth-resource-provider

Change-Id: Ia2e6a65690a1d7f89af0bc32fa8be30ca09472e2