Make supports_direct_io work on 4096b sector size
The current check uses an alignment of 512 bytes and will fail when the
underlying device has sectors of size 4096 bytes, as is common e.g. for
NVMe disks. So use an alignment of 4096 bytes, which is a multiple of
512 bytes and thus will cover both cases.

Conflicts:
        nova/privsep/utils.py
        - supports_direct_io() is still in
          nova/virt/libvirt/driver.py for this branch

Change-Id: I5151ae01e90506747860d9780547b0d4ce91d8bc
Closes-Bug: 1801702
Co-Authored-By: Alexandre Arents <alexandre.arents@corp.ovh.com>
(cherry picked from commit 14d98ef1b48ca7b2ea468a8f1ec967b954955a63)