Improve libvirt image and snapshot handling
Use virt.images.convert_image instead of invoking 'qemu-img convert'
directly so that cache option can be set properly. i.e. cache=none
when O_DIRECT is supported by the filesystem.

The big advantage of this is that it will bypass the host page cache
so that converting large images won't evict guest data from the cache.

Refactor images.convert_image so that the compression option can be
controlled by the caller.

Change-Id: I4b7be98b5832ca8c580339fcfb7b9203264b5ff8
Signed-off-by: Jack Ding <jack.ding@windriver.com>
Signed-off-by: Chris Friesen <chris.friesen@windriver.com>