libvirt: change "Ignoring supplied device name" warning to info
This doesn't need to be a warning that only the operator sees
in the nova-compute logs. The operator can't do anything about
it since the REST API user is the one passing in a device name
when attaching a volume to their server. So let's just change
the warning to an info level message. We already have documentation
stating that the device parameter is ignored by some compute
drivers.

Change-Id: If2dd8d0a3992ccf4bd9aa01978b7c91897e2eb59