Delete NeutronLinuxBridgeInterfaceDriver
This is a really old carryover from when neutron was
quantum and the virt drivers would plug/unplug using
the driver defined by the linuxnet_interface_driver
config option, which is deprecated and only used for
nova-network now. We use os-vif now for plugging and
unplugging bridge VIFs with neutron so this old
unused driver code can be removed. It wasn't even
tested.

Change-Id: I07f25685e4fb650c22838bc3afab2bb40809e438
Related-Bug: #1801919