Fill the RequestGroup mapping during schedule
Use the previously introduced map_requested_resources_to_providers()
function in the RequestSpec during the scheduling of a new server. This
is done in the conductor becase later when placement will return such mapping
this will be the place when that mapping can be moved form the Selection
object returned by the scheduler to the RequestSpec.

blueprint bandwidth-resource-provider

Change-Id: I0c764e441993e32aafef0b18049a425c3c832a50