Add functional regression recreate test for bug 1790204
Since Pike, the FilterScheduler "claims" resources in
placement via allocations based on the flavor during
scheduling. For a same-host resize, the old and new flavor
are summed to get the overall allocation against the single
compute node resource provider, which can cause NoValidHost
failures during scheduling because the sum is over capacity
for the resource provider even though the new flavor alone
may be OK.

This adds a functional regression test to recreate the bug.

Change-Id: I036a5ceabe88dcc1fd85c09472481de7d02edf5f
Related-Bug: #1790204