Pass resource provider mapping to neutronv2 api
The RequestSpec object contains a mapping between request groups and
resource providers fullfilling the request of the group. This patch
sends this mapping down to neutronv2 api so that the resource provider
UUID can be included in the port binding if the port requested resources.

The next patch will modify the port binding logic to use this
information.

blueprint bandwidth-resource-provider

Change-Id: Ie5ea45484a632da77c41bfe2613fa5f1dfbf19ea