Drop cruft code for all_tenants behaviour
The behaviour for the Nova API is to require `all_tenants` to be
included in order to be able to list instances for a specific
project.

While this behaviour can be a bit confusing, all official CLI
clients currently automatically imply the addition of `all_tenants`
to the API request if an explicit tenant/project is selected when
listing instances.

The TODO has lived in the code for 5 years now without any work
being done on it.  This patch drops the commented code which can
later be updated with the appropriate microversion bump to allow
for behaviour that doesn't require including all_tenants in the
HTTP request for the future.

Related-Bug: #1185290

Change-Id: I9324f798b3c6e05991802bb6610fc357864c0c66