Provide a useful error message when trying to update non-compute services
Starting in Pike, we disallowed trying to update (enable/disable/force down)
non-nova-compute services because of multi-cell support using host mappings
to lookup service records, and simply because disabling non-compute services
doesn't do anything.

However, before microversion 2.53, the error the user gets back is confusing:

  HTTP exception thrown: Host 'p024.domain.com' is not mapped to any cell

This change provides a useful error message in this case and also changes
the 404 response to a 400 response to align with the type of error and the
behavior of the 2.53 microversion.

Change-Id: I44f09aec60b0b18c458f9ba6d8b725db962e9cc7
Closes-Bug: #1805164