Mention size limit on user data in docs
The API reference for the "user_data" parameter to server
create and rebuild mentions the size restriction but the
user docs did not, so this adds that note. While doing this,
several other user-facing docs that mention user data link back
to the overall user-data doc so that we can centralize any
further documentation about that topic in a single location.

Change-Id: Id4a61d58150337e0dec223d4d2741336ed6d5387
Closes-Bug: #1804653