Make [cinder]/catalog_info no longer require a service_name
The service_name part of the cinder catalog_info option is not
necessary to lookup the endpoint from the service catalog when
we have the endpoint type (volumev3) and the interface (publicURL).

This changes the option default to not include the service name
and no longer makes a service name required. If one is provided,
then it will be sent along to KSA/cinderclient, otherwise it is
omitted.

As this is a change in behavior of the config option, a release note
is added.

Change-Id: I89395fafffd60981fba17a7b09f7015e1f827b62
Closes-Bug: #1803627