Drop pre-cinder 3.44 version compatibility
The new style volume attach flow semantics, which
were needed to support multiattach volumes, relies
on cinder v3.44 and was added in Queens:

  Ifc01dbf98545104c998ab96f65ff8623a6db0f28

Now that we're in Train, let's require Cinder API
to be at least running Queens level code and drop
the compat checks for queens level nova-compute code
and Cinder API older than Queens.

Lots of old style attachment test code is dropped
as a result since we don't need the version boundary
handling in the API code. The way a lot of the "new
flow" tests were written were as copies of existing
tests modified for the new flow so the old flow tests
could just be dropped later, and that's what happens
in this change.

Change-Id: I6a777b4b7a5729488f939df8c40e49bd40aec3dd