Handle tags in _bury_in_cell0
We allowed add instance tags when booting in
Iac54b9627cb4398e45f1f15a2f4e7d6f86242093, conductor
will create tags for us in the selected cell, but
we forgot to consider the case that no valid host
was found.

So we will be unable to filter those instances in
cell0 with tags.

Change-Id: Idfe37d356b349024dfe963c934728111e1a1d314
closes-bug: #1806515