Only warn about not having computes nodes once in rpcapi
When we instantiate the compute rpcapi, with upgrade_levels=auto,
we determine the minimum service version of nova-compute. If we
find no computes, we log a message. This may be known, if we are
starting at zero day and no computes have started, but may also
indicate a more serious problem. In either case, logging this once
per process (especially in the case of multi-worker, multi-module
services) is more than enough.

Change-Id: I3350026c872a548839b857c53529c1dbbb88091c
Related-Bug: #1807044