Make compute rpcapi version calculation check all cells
For top-level services, the compute rpcapi module needs to check for the
minimum version of nova-compute in all cells in order to function
properly during an upgrade. This moves that module to use the all-cells
variant of this routine in the case where [api_database]/connection is
set.

Depends-On: https://review.openstack.org/625039/

Change-Id: Icddbe4760eaff30e4e13c1e8d3d5d3f489dac3c4
Closes-Bug: #1807044
(cherry picked from commit d779b33d72d6ef41651ce7b93fe982f121bae2d7)