Create BDMs/tags in cell with instance when over-quota
If the server create build request fails the quota check
after the instance record has been created in a cell, we also
need to create the BDMs and tags in that cell so that users
can still see the tags on the server and so the API can
properly cleanup volume attachments when the server is deleted.

This change updates _cleanup_build_artifacts to create BDMs
and tags in the same cell as the instance prior to deleting the
build request and request spec and adjusts the assertions in the
related functional test to show the bug is fixed.

As for instances that get buried in cell0 due to scheduling
failures, the tags are not created there so comments are left
in those code paths to fix that issue as well, but that can be
done separately from this patch.

NOTE(s10): Conflict is caused by I70b11dd489d222be3d70733355bfe7966df556aa
not being in Pike.

Change-Id: I1a9bdb596f74605ab4613c9cb2574e976aebbd8c
Closes-Bug: #1806064
(cherry picked from commit 6d0386058b9628bbfcf64abdd707ad87ee19353c)
(cherry picked from commit 3028d25705bbd54cdf0b7ba13859a809d401fc70)
(cherry picked from commit 3dd42c2cd658fd3f73b11fbf5e81ccccd748b450)