Don't call begin_detaching when detaching volume from shelved vm
When shelve an instance, if the instance has volume attached,
with new attach/detach flow, we will delete the old attachment
and create a new attachment, the volume status will be ``reserved``.

If the user tries to detach these volumes, it fails due to that
Cinder does not allow a begin_detaching() call on a `reserved` volume.

Actually for shelved instances, we can just skip this step and
directly detach it.

Change-Id: Ib1799feebbd8f4b0f389168939df7e5e90c8add1
closes-bug: #1808089
(cherry picked from commit 41b982c9feec3105247bc72d23d1470bcabc3a0f)